Raise ParseException in yangParser.parse()
