+++ utility/util.py
@@ -106,18 +106,20 @@ def find_first_file(directory: str, pattern: str,
         for pattern in patterns_order:
             filename = match_file(path, pattern)
             if filename:
-                parsed = yangParser.parse(filename)
-                if parsed:
-                    results = parsed.search('revision')
-                    if results:
-                        revision = results[0].arg
-                    else:
-                        revision = '1970-01-01'
-                    if '*' not in pattern_with_revision:
-                        if revision in pattern_with_revision:
-                            return filename
-                    else:
+                try:
+                    parsed = yangParser.parse(filename)
+                except:
+                    continue
+                results = parsed.search('revision')
+                if results:
+                    revision = results[0].arg
+                else:
+                    revision = '1970-01-01'
+                if '*' not in pattern_with_revision:
+                    if revision in pattern_with_revision:
                         return filename
+                else:
+                    return filename
 
 
 def change_permissions_recursive(path: str):
